+++ projects/frontend/es-app/src/emulators/run/GameRunner.cpp
@@ -183,11 +183,17 @@ bool GameRunner::RunGame(FileData& game, const EmulatorData& emulator, const Gam
     Board::Instance().SetCPUGovernance(GetGovernance(core));
     Board::Instance().StartInGameBackgroundProcesses(sdl2Runner);
     fputs("==============================================\n", stdout);
+    long startGameTime = DateTime().ToEpochTime();
+
     // Start game thread
     ThreadRunner gameRunner(sdl2Runner, command, debug);
     // Start sdl2 loop
     sdl2Runner.Run();
     exitCode = gameRunner.ExitCode();
+
+    long stopGameTime = DateTime().ToEpochTime();
+
+
     fputs("==============================================\n", stdout);
     Board::Instance().StopInGameBackgroundProcesses(sdl2Runner);
     Board::Instance().SetFrontendCPUGovernor();
@@ -195,7 +201,12 @@ bool GameRunner::RunGame(FileData& game, const EmulatorData& emulator, const Gam
     padToKeyboard.StopMapping();
 
     if (exitCode != 0) { LOG(LogWarning) << "[Run] Non-Zero exit code " << exitCode << " !"; }
-    else LOG(LogInfo) << "[Run] No error running " << path.ToString();
+    else
+    {
+      LOG(LogInfo) << "[Run] No error running " << path.ToString();
+      // Update time played
+      game.Metadata().SetTimePlayed(game.Metadata().TimePlayed() + (int) stopGameTime - (int) startGameTime);
+    }
   }
 
   SubSystemRestore();